Check for OpenSSL Heartbleed bug at build time
To facilitate this, add ability to query version information for the OpenSSL
library being used in the build. Add a test that fails if the library being
linked against is vulnerable. User can still force install, but this
prevents automated install on a system with vulnerable OpenSSL. Especially
important since Crypt::SSLeay does not verify hosts.