Use a define'd constant rather than sizeof
In `SSL_CTX_new`, if loading from `/dev/urandom` fails, the contents of a
local char array is used as a seed source. The size of this memory block is
passed using sizeof, but if the argument was ever changed to be `char *`,
this would not be good. So, instead, use a define'd constant.