Also consider $ENV{OPENSSL_PREFIX} and --openssl-prefix
Net::SSLeay defers to $ENV{OPENSSL_PREFIX} if it is defined. If a user
has that setting, there is no need to add specific environment variables just
for our sake. So, $ENV{OPENSSL_PREFIX} exists, then we'll look for OpenSSL
headers in $ENV{OPENSSL_PREFIX}/include and libs in $ENV{OPENSSL_PREFIX}/lib.