Makefile.PL: More carefully and precisely load from inc/
This more carefully loads the dependency from `inc` without
necessitating re-polluting `@INC` globally.

This also negates various bugs in the existing implementation,
for instance, the fact that `require` auto-fatalizes, and the fact
that in order to avoid `@INC` traversal in conjunction with `require`,
you need to specify a path prefixed with either `./` or `/`.

`catfile` is useless here, as things like:

```
catfile(qw(. foo bar baz ))
```

Will helpfully eat the "."

Additionally, `require` by convention takes paths in Unix form, and
using windows paths to require should be considered a bug.