Use ExtUtils::PkgConfig to discover OpenSSL if available (#8)
Makefile.PL used to link to zlib regardless OpenSSL libraries needed
that or not. Just based on the presente of zlib on the system.

If OpenSSL required zlib but it was not installed, Crypt-SSLeay tests
failed because of underlinking. If OpenSSL did not require zlib, but
it was installed, Crypt/SSLeay.so pulled in useless zlib because of
overlinking. Overlinking polutes a process address space, enlarges an
attack surface and slows down execution.

This patch modifies Makefile.PL to optionally use ExtUtils::PkgConfig
to discover OpenSSL and its compiler and linker flags.